% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arboles}
\alias{arboles}
\title{Caracteristicas de arboles}
\format{
Data con 90 registros y 4 variables
\describe{
\item{id}{Identificador único del arbol}
\item{peso}{Peso del árbol en kilogramos}
\item{diametro}{Diamétro del árbol en metros}
\item{altura}{Altura del árbol en metros}
}
}
\usage{
  arboles
}
\description{
  Esta data contiene caracteristicas de arboles
}
\examples{
  data(arboles)
}
\keyword{datasets}
