% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arboles1}
\alias{arboles1}
\title{Caracteristicas de arboles}
\format{
Data con 90 registros y 5 variables
\describe{
\item{finca}{Finca a la que pertenece el árbol}
\item{mg}{Genotipo que posee el arbol}
\item{peso}{Peso del árbol en kilogramos}
\item{diametro}{Diamétro del árbol en metros}
\item{altura}{Altura del árbol en metros}
}
}
\usage{
  arboles1
}
\description{
  Esta data contiene caracteristicas de arboles
}
\examples{
  data(arboles1)
}
\keyword{datasets}
