% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{corazon}
\alias{corazon}
\title{Caracteristicas de pacientes}
\format{
Data con 50 registros y 6 variables
\describe{
\item{id}{identificacion paciente}
\item{edad}{edad del paciente en años}
\item{actividad}{tipo de actividad deportiva que realiza : 0 = ninguna, 1= mínima, 2=moderada, 3=mas que moderada}
\item{angina}{ocurrencia o no de angina de pecho : 1: si, 0: no}
\item{presion}{si presenta presión sanginea alta =1, y 0 si no la pesenta}
\item{infarto}{si ha presentado infarto en el miocardio 1, y 0 si no ha presentado}
}
}
\usage{
  corazon
}
\description{
  Esta data contiene caracteristicas depacientes, tomada de Métodos avanzados de estadística aplicada. Alfonso Garcia Perez
}
\examples{
  data(corazon)
}
\keyword{datasets}
