% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{creditos}
\alias{creditos}
\title{Morosidad en creditos}
\format{
Data con 780 registros y 5 variables
\describe{
\item{defalut}{Indica si el cliente ha sido moroso, toma el valor de 1 si lo ha sido.}
\item{antiguedad}{Años que lleva el cliente en la entidad.}
\item{edad}{Edad del cliente.}
\item{cuota}{Cantidad de dinero en pesos colombianos que pidió prestado el cliente.}
\item{ingresos}{Ingresos mensuales del cliente.}
}
}
\usage{
  creditos
}
\description{
  Esta data contiene información sobre clientes morosos y no morosos de una entidad financiera.
}
\examples{
  data(creditos)
}
\keyword{datasets}