% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eleccion}
\alias{eleccion}
\title{Eleccion de la carrera en ingenieria}
\format{
Data con 200 registros y 2 variables
\describe{
\item{matricula}{Indicador si el estudiante elige una carrera en ingeniería}
\item{matematicas}{Puntaje obtenido en matemáticas}
}
}
\usage{
  eleccion
}
\description{
  Esta data contiene información sobre estudiantes que ingresan a la universidad.
}
\examples{
  data(eleccion)
}
\keyword{datasets}