% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rotacion}
\alias{rotacion}
\title{Empleados de compañía}
\format{
Data con 1470 registros y 24 variables
\describe{
\item{Rotación}{Indica si el trabajador ha sido cambiado de departamento.}
\item{Edad}{Edad del empleado.}
\item{Viaje de negocios}{Indica con que frecuencia el empleado viaja debido a negocios.}
\item{Departamento}{Departamento a la que pertenece.}
\item{Distancia_Casa}{Distancia en kilometros de la casa hasta la oficina.}
\item{Educación}{Nivel educativo.}
\item{Campo_Educación}{Campo educativo.}
\item{Satisfacción_Ambiental}{Percepción ambiental del empleado.}
\item{Genero}{Genero.}
\item{Cargo}{Cargo que ocupa actualmente}
\item{Satisfación_Laboral}{Percepción laboral del empleado.}
\item{Estado_Civil}{Estado civil}
\item{Ingreso_Mensual}{Ingreso mensual en dólares.}
\item{Trabajos_Anteriores}{Número de trabajos anteriores que tuvo.}
\item{Horas_Extra}{Indica trabaja horas extra.}
\item{Porcentaje_aumento_salarial}{Aumento salario del cargo actual respecto al anterior.}
\item{Rendimiento_Laboral}{Puntaje sobre el rendimiento laboral del empleado.}
\item{Años_Experiencia}{Años de experiencia.}
\item{Capacitaciones}{Número de capacitaciones.}
\item{Equilibrio_Trabajo_Vida}{Puntaje que da el empleado sobre el equilibrio entre el trabajo y su vida personal.}
\item{Antigüedad}{Años en la empresa}
\item{Antigüedad_Cargo}{Años que ocupa el empleado en su actual cargo.}
\item{Años_ultima_promoción}{Años que han pasado desde el último ascenso.}
\item{Años_acargo_con_mismo_jefe}{Años con el mismo jefe de área.}
}
}
\usage{
  rotacion
}
\description{
  Esta data contiene información sobre trabajadores de una compañía.
}
\examples{
  data(rotacion)
}
\keyword{datasets}